Fixes #7232. Fixes problem with enum handling in latest gcc.
git-svn-id: http://svn.mythtv.org/svn/trunk@22254 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/hdhrchannel.cpp
@@ -60,7 +60,7 @@ bool HDHRChannel::Open(void)
 
     _tuner_types = _stream_handler->GetTunerTypes();
     tunerType = (_tuner_types.empty()) ?
-        DTVTunerType::kTunerTypeUnknown : _tuner_types[0];
+        DTVTunerType::kTunerTypeUnknown : (int) _tuner_types[0];
 
     if (!InitializeInputs())
     {
